legacy_backup/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
anthropic==0.18.1
asyncpg==0.29.0
redis==5.0.1
